v2.9.07: Added "C" method of FIFO expanding to 5 elements, first test appers to work properly.

+++ EBB_firmware/app.X/source/UBW.h
@@ -143,7 +143,7 @@ extern unsigned char error_byte;
 //extern unsigned char error_byte;
 extern BOOL	g_ack_enable;
 
-extern volatile unsigned int g_RC_value[kRC_DATA_SIZE];   // Stores reload values for TMR0
+extern volatile UINT16 g_RC_value[kRC_DATA_SIZE];   // Stores reload values for TMR0
 extern volatile tRC_state g_RC_state[kRC_DATA_SIZE];
 
 extern volatile unsigned long int gRCServoPoweroffCounterMS;
